landscape page's content is sorted
