+++ structurizr-gen/sample/md/README.md
@@ -6,8 +6,8 @@
 
 ## Software Systems
 
- - [Spring PetClinic](Spring%20PetClinic.md)
  - [App 2](App%202.md)
+ - [Spring PetClinic](Spring%20PetClinic.md)
 
 ## Stats
 
@@ -15,36 +15,36 @@
 
 | Software System | # Containers | # Components |
 | --- | --- | --- |
-| Spring PetClinic | 2 | 2 |
 | App 2 | 2 | 2 |
+| Spring PetClinic | 2 | 2 |
 
 ### Containers
 
 | Software System | Container | # Components |
 | --- | --- | --- |
-| Spring PetClinic | Web Application | 2 |
-| Spring PetClinic | Database | 0 |
-| App 2 | Web Application | 2 |
 | App 2 | Database | 0 |
+| App 2 | Web Application | 2 |
+| Spring PetClinic | Database | 0 |
+| Spring PetClinic | Web Application | 2 |
 
 ## Relations
 
 | Software System | # |
 | --- | --- |
-| Spring PetClinic |  |  | 0 |
 | App 2 |  |  | 5 |
+| Spring PetClinic |  |  | 0 |
 
 | Software System | Container | # |
 | --- | --- | --- |
-| Spring PetClinic | Web Application |  | 1 |
-| Spring PetClinic | Database |  | 0 |
-| App 2 | Web Application |  | 5 |
 | App 2 | Database |  | 0 |
+| App 2 | Web Application |  | 5 |
+| Spring PetClinic | Database |  | 0 |
+| Spring PetClinic | Web Application |  | 1 |
 
 | Software System | Container | Component | # |
 | --- | --- | --- | --- |
-| Spring PetClinic | Web Application | Web Service 1 | 1 |
-| Spring PetClinic | Web Application | Web Service 2 | 1 |
 | App 2 | Web Application | Web Service 3 | 3 |
 | App 2 | Web Application | Web Service 4 | 4 |
+| Spring PetClinic | Web Application | Web Service 1 | 1 |
+| Spring PetClinic | Web Application | Web Service 2 | 1 |
